requirements.txt
Flask
requests
python-dotenv
google-generativeai
flask_cors
